test/requirements.txt
# Веб-фреймворк и сервер
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.3
python-multipart==0.0.6

# Векторная база данных и эмбеддинги
chromadb==0.4.18
sentence-transformers==2.2.2

torch==2.2.0
numpy==1.24.3
transformers==4.35.2
accelerate==0.24.1
tokenizers==0.15.0

# Клиент для LMStudio
openai==1.3.9
requests==2.31.0

# Обработка документов
pypdf==3.17.0
python-docx==1.1.0
unstructured==0.10.30
langchain-text-splitters==0.0.1

aiofiles==23.2.1
python-dotenv==1.0.0
loguru==0.7.2
psutil==5.9.6
tiktoken==0.5.2

python-jose==3.3.0
passlib[bcrypt]==1.7.4

pydantic-settings==2.0.3